+++ lib/src/jwt/dart_jsonwebtoken.dart
@@ -1,7 +1,0 @@
-library jsonwebtoken;
-
-export 'src/jwt.dart';
-export 'src/errors.dart';
-export 'src/algorithms.dart';
-export 'src/keys.dart';
-export 'src/parser.dart';