+++ lib/src/rsa/rsa_api.dart
@@ -91,7 +91,7 @@ class GeneratedRsaKeyPair {
 /// ## Usage
 ///
 /// ```dart
-/// import 'package:solid_auth/src/rsa/rsa_impl.dart';
+/// import 'package:solid_oidc_auth/src/rsa/rsa_impl.dart';
 ///
 /// // Use the singleton instance
 /// final result = await rsa.generate(2048);